# Accelerated Merkle Tree Construction using Rescue Prime