--- a/Makefile
+++ b/Makefile
@@ -60,6 +60,12 @@ format:
 tests/ff_p.o: ff_p.cpp
 	$(CXX) $(CXXFLAGS) $(SYCLFLAGS) $(DFLAGS) -c $^ -o $@ $(INCLUDES)
 
+tests/merkle_tree.o: merkle_tree.cpp
+	$(CXX) $(CXXFLAGS) $(SYCLFLAGS) $(DFLAGS) -c $^ -o $@ $(INCLUDES)
+
+tests/test_merkle_tree.o: tests/test_merkle_tree.cpp
+	$(CXX) $(CXXFLAGS) $(SYCLFLAGS) $(DFLAGS) -c $^ -o $@ $(INCLUDES)
+
 tests/rescue_prime.o: rescue_prime.cpp
 	$(CXX) $(CXXFLAGS) $(SYCLFLAGS) $(DFLAGS) -c $^ -o $@ $(INCLUDES)
 
@@ -78,7 +84,7 @@ tests/test.o:	tests/test.cpp
 tests/main.o: tests/main.cpp
 	$(CXX) $(CXXFLAGS) $(SYCLFLAGS) $(DFLAGS) -c $^ -o $@ $(INCLUDES)
 
-test: tests/ff_p.o tests/rescue_prime.o tests/test_rescue_prime.o tests/ntt.o tests/test_ntt.o tests/test.o tests/main.o
+test: tests/ff_p.o tests/merkle_tree.o tests/test_merkle_tree.o tests/rescue_prime.o tests/test_rescue_prime.o tests/ntt.o tests/test_ntt.o tests/test.o tests/main.o
 	$(CXX) $(SYCLFLAGS) $^ -o tests/$@
 	@./tests/$@