--- a/include/bench_merkle_tree.hpp
+++ b/include/bench_merkle_tree.hpp
@@ -1,0 +1,10 @@
+#pragma once
+#include <CL/sycl.hpp>
+
+// Returns exact time of kernel execution in nanosecond
+//
+// Ensure SYCL queue has profiling enabled !
+uint64_t
+benchmark_merklize(sycl::queue& q,
+                   const size_t leaf_count,
+                   const size_t wg_size);