--- a/include/bench_merkle_tree.hpp
+++ b/include/bench_merkle_tree.hpp
@@ -5,6 +5,14 @@
 //
 // Ensure SYCL queue has profiling enabled !
 uint64_t
-benchmark_merklize(sycl::queue& q,
-                   const size_t leaf_count,
-                   const size_t wg_size);
+benchmark_merklize_approach_1(sycl::queue& q,
+                              const size_t leaf_count,
+                              const size_t wg_size);
+
+// Returns sum of all kernel execution times in nanosecond
+//
+// Ensure SYCL queue has profiling enabled !
+uint64_t
+benchmark_merklize_approach_2(sycl::queue& q,
+                              const size_t leaf_count,
+                              const size_t wg_size);