--- a/include/merkle_tree.hpp
+++ b/include/merkle_tree.hpp
@@ -11,11 +11,38 @@
 // Returns sum of all kernel execution times with nanosecond
 // level granularity
 uint64_t
-merklize(sycl::queue& q,
-         const sycl::ulong* leaves,
-         sycl::ulong* const intermediates,
-         const size_t leaf_count,
-         const size_t wg_size,
-         const sycl::ulong16* mds,
-         const sycl::ulong16* ark1,
-         const sycl::ulong16* ark2);
+merklize_approach_1(sycl::queue& q,
+                    const sycl::ulong* leaves,
+                    sycl::ulong* const intermediates,
+                    const size_t leaf_count,
+                    const size_t wg_size,
+                    const sycl::ulong16* mds,
+                    const sycl::ulong16* ark1,
+                    const sycl::ulong16* ark2);
+
+// Same as above routine, serves similar purpose, when
+// N -many leaves of binary merkle tree are provided, computes
+// all N-1 -many intermediates, while using Rescue Prime merge function
+// for merging two digests, which is equivalent to computing parent node
+// from two children
+//
+// N -needs to be power of 2
+//
+// Check those assertions, I've written in implementation of this routine
+// which are my assumptions, feel free to tinker with them !
+//
+// Returns total time spent on computing merkle tree intermediate nodes
+// with nanosecond level granularity
+//
+// Have taken major motivation from
+// https://github.com/itzmeanjan/vectorized-rescue-prime/blob/c48b8555e07eb9557a20383cc9f3a4aeec834317/rescue_prime.c#L153-L164
+// where I wrote similar routine using OpenCL
+uint64_t
+merklize_approach_2(sycl::queue& q,
+                    const sycl::ulong* leaves,
+                    sycl::ulong* const intermediates,
+                    const size_t leaf_count,
+                    const size_t wg_size,
+                    const sycl::ulong16* mds,
+                    const sycl::ulong16* ark1,
+                    const sycl::ulong16* ark2);