--- a/include/test_merkle_tree.hpp
+++ b/include/test_merkle_tree.hpp
@@ -1,0 +1,8 @@
+#pragma once
+#include <merkle_tree.hpp>
+#include <random>
+#include <cassert>
+
+// Test that merkle tree construction kernel is working as expected !
+void
+test_merklize(sycl::queue& q);