--- a/main.cpp
+++ b/main.cpp
@@ -86,7 +86,7 @@ main(int argc, char** argv)
 
   for (uint dim = (1ul << 20); dim <= (1ul << 23); dim <<= 1) {
     // time in nanoseconds --- beware !
-    uint64_t tm = benchmark_merklize(q, dim, 1ul << 6);
+    uint64_t tm = benchmark_merklize(q, dim, 1ul << 5);
 
     std::cout << std::setw(11) << std::right << dim << "\t\t" << std::setw(15)
               << std::right << tm * 1e-6 << " ms" << std::endl;