--- a/main.cpp
+++ b/main.cpp
@@ -78,15 +78,29 @@ main(int argc, char** argv)
               << 1e9 / ((double)tm / (double)(dim * dim * 1)) << std::endl;
   }
 
-  std::cout
-    << "\nMerklize using Rescue Prime on F(2**64 - 2**32 + 1) elements 👇\n"
-    << std::endl;
+  std::cout << "\nMerklize ( approach 1 ) using Rescue Prime on F(2**64 - "
+               "2**32 + 1) elements 👇\n"
+            << std::endl;
   std::cout << std::setw(11) << "dimension"
             << "\t\t" << std::setw(15) << "total" << std::endl;
 
   for (uint dim = (1ul << 20); dim <= (1ul << 23); dim <<= 1) {
     // time in nanoseconds --- beware !
-    uint64_t tm = benchmark_merklize(q, dim, 1ul << 5);
+    uint64_t tm = benchmark_merklize_approach_1(q, dim, 1ul << 5);
+
+    std::cout << std::setw(11) << std::right << dim << "\t\t" << std::setw(15)
+              << std::right << tm * 1e-6 << " ms" << std::endl;
+  }
+
+  std::cout << "\nMerklize ( approach 2 ) using Rescue Prime on F(2**64 - "
+               "2**32 + 1) elements 👇\n"
+            << std::endl;
+  std::cout << std::setw(11) << "dimension"
+            << "\t\t" << std::setw(15) << "total" << std::endl;
+
+  for (uint dim = (1ul << 20); dim <= (1ul << 23); dim <<= 1) {
+    // time in nanoseconds --- beware !
+    uint64_t tm = benchmark_merklize_approach_2(q, dim, 1ul << 5);
 
     std::cout << std::setw(11) << std::right << dim << "\t\t" << std::setw(15)
               << std::right << tm * 1e-6 << " ms" << std::endl;