… tests/test_merkle_tree.cpp