This PR implements routines for parallelizing computation of intermediate nodes of fully balanced Binary Merkle Tree, while taking some inspiration from [here](https://github.com/itzmeanjan/vectorized-rescue-prime/blob/77e371ef2fb11ba7d7369005a60a0888393729f0/rescue_prime.c#L153-L164), which I wrote in OpenCL, sometime ago. 

Now I've written Merkle Tree kernel dispatching/ orchestration logic in two different ways and here are respective benchmark results when I compiled using [Intel's SYCL compiler](https://intel.github.io/llvm-docs/GetStartedGuide.html) using CUDA backend on Nvidia Tesla V100.

 Leaf Count | Approach 1 | Approach 2
--- | --- | ---
1048576 | 180.584 ms | 89.2852 ms
2097152 | 296.861 ms | 162.189 ms
4194304 | 545.38 ms | 306.154 ms
8388608 | 1065.68 ms | 593.262 ms

> Note approach 2 is nothing but SYCL translation of OpenCL kernel over [here](https://github.com/itzmeanjan/vectorized-rescue-prime/blob/c48b8555e07eb9557a20383cc9f3a4aeec834317/rescue_prime.c#L153-L164)

> You'll be probably interested in looking at [this](https://github.com/itzmeanjan/vectorized-rescue-prime/tree/77e371ef2fb11ba7d7369005a60a0888393729f0#nvidia-tesla-v100-gpu-with-opencl) benchmark obtained on same GPU, when aforementioned approach 2 was written in OpenCL. 

I plan to do deeper analysis of both of these techniques.